+++ lib/permisi/backend/active_record.rb
@@ -1,22 +1,26 @@
+# frozen_string_literal: true
+
 require "active_record"
 
 module Permisi
-  class Backend::ActiveRecord < Backend::Base
-    class << self
-      def table_name_prefix
-        "permisi_"
-      end
+  module Backend
+    module ActiveRecord
+      class << self
+        def table_name_prefix
+          "permisi_"
+        end
 
-      def findsert_actor(aka)
-        Actor.find_or_create_by(aka: aka)
-      end
+        def findsert_actor(aka)
+          Actor.find_or_create_by(aka: aka)
+        end
 
-      def actors
-        Actor.all
-      end
+        def actors
+          Actor.all
+        end
 
-      def roles
-        Role.all
+        def roles
+          Role.all
+        end
       end
     end
   end