+++ lib/permisi/backend/active_record/actor_role.rb
@@ -1,6 +1,20 @@
+# frozen_string_literal: true
+
 module Permisi
-  class Backend::ActiveRecord::ActorRole < ::ActiveRecord::Base
-    belongs_to :actor
-    belongs_to :role
+  module Backend
+    module ActiveRecord
+      class ActorRole < ::ActiveRecord::Base
+        belongs_to :actor, touch: true
+        belongs_to :role
+
+        after_destroy :touch_actor
+
+        private
+
+        def touch_actor
+          actor.touch
+        end
+      end
+    end
   end
 end