+++ spec/support/user.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
-module Permisi
-  VERSION = "0.1.1"
+class User < ActiveRecord::Base
 end