+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 0.1.1
+
+- General code refactoring
+- Improvements on ActiveRecord backend:
+  - Code refactoring
+  - Implement cache invalidation
+
 # 0.1.0
 
 Finished extraction work from my past projects.